Fix changes requested

+++ src/Repository/WishlistRepository.php
@@ -21,7 +21,7 @@ class WishlistRepository
 {
     public function getAllWishlistsProductID()
     {
-        return Db::getInstance()
+        return (int) Db::getInstance()
             ->getRow('SELECT `id_product` FROM `' . _DB_PREFIX_ . 'wishlist_product`');
     }
 }
